# api-gateway-service/main.go
package main

import (
	"context"
	"log"
	"net/http"
	"os"
	"os/signal"
	"syscall"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/spf13/viper"
)

func main() {
	// Load configuration
	if err := loadConfig(); err != nil {
		log.Fatalf("Failed to load configuration: %v", err)
	}

	// Initialize router
	router := setupRouter()

	// Configure server
	server := &http.Server{
		Addr:         viper.GetString("server.address"),
		Handler:      router,
		ReadTimeout:  viper.GetDuration("server.read_timeout"),
		WriteTimeout: viper.GetDuration("server.write_timeout"),
	}

	// Start server in a goroutine
	go func() {
		if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
			log.Fatalf("Failed to start server: %v", err)
		}
	}()

	// Wait for interrupt signal
	quit := make(chan os.Signal, 1)
	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
	<-quit
	log.Println("Shutting down server...")

	// Graceful shutdown
	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()
	if err := server.Shutdown(ctx); err != nil {
		log.Fatalf("Server forced to shutdown: %v", err)
	}

	log.Println("Server exited")
}

func loadConfig() error {
	viper.SetConfigName("config")
	viper.SetConfigType("yaml")
	viper.AddConfigPath(".")
	viper.AddConfigPath("./config")

	// Set defaults
	viper.SetDefault("server.address", ":8080")
	viper.SetDefault("server.read_timeout", 10*time.Second)
	viper.SetDefault("server.write_timeout", 10*time.Second)
	viper.SetDefault("rate_limit.enabled", true)
	viper.SetDefault("rate_limit.requests_per_second", 10)
	viper.SetDefault("auth.jwt_secret", "")
	viper.SetDefault("auth.token_expiry", 24*time.Hour)

	if err := viper.ReadInConfig(); err != nil {
		if _, ok := err.(viper.ConfigFileNotFoundError); !ok {
			return err
		}
	}

	return nil
}

func setupRouter() *gin.Engine {
	if !viper.GetBool("debug") {
		gin.SetMode(gin.ReleaseMode)
	}

	router := gin.New()
	router.Use(gin.Recovery())

	// Middleware
	router.Use(corsMiddleware())
	router.Use(loggerMiddleware())
	if viper.GetBool("rate_limit.enabled") {
		router.Use(rateLimitMiddleware())
	}

	// Health check
	router.GET("/health", healthCheck)

	// API routes
	api := router.Group("/api/v1")
	api.Use(authMiddleware())
	{
		// Cost analysis endpoints
		costs := api.Group("/costs")
		{
			costs.GET("", getCosts)
			costs.GET("/summary", getCostSummary)
			costs.GET("/forecast", getCostForecast)
		}

		// Resource optimization endpoints
		optimize := api.Group("/optimize")
		{
			optimize.POST("/analyze", analyzeResources)
			optimize.GET("/recommendations", getRecommendations)
			optimize.POST("/apply", applyRecommendations)
		}

		// Provider management endpoints
		providers := api.Group("/providers")
		{
			providers.GET("", getProviders)
			providers.GET("/:provider", getProviderDetails)
			providers.POST("/:provider/connect", connectProvider)
			providers.DELETE("/:provider/disconnect", disconnectProvider)
		}

		// Resource management endpoints
		resources := api.Group("/resources")
		{
			resources.GET("", getResources)
			resources.GET("/:id", getResource)
			resources.POST("/scan", scanResources)
			resources.POST("/tag", tagResources)
		}
	}

	return router
}

// Middleware implementations
func corsMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		c.Header("Access-Control-Allow-Origin", "*")
		c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Authorization")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	}
}

func loggerMiddleware() gin.HandlerFunc {
	return gin.LoggerWithConfig(gin.LoggerConfig{
		SkipPaths: []string{"/health"},
	})
}

func rateLimitMiddleware() gin.HandlerFunc {
	// TODO: Implement rate limiting
	return func(c *gin.Context) {
		c.Next()
	}
}

func authMiddleware() gin.HandlerFunc {
	// TODO: Implement JWT authentication
	return func(c *gin.Context) {
		c.Next()
	}
}

// Handler implementations
func healthCheck(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"status": "healthy",
		"time":   time.Now().UTC(),
	})
}

func getCosts(c *gin.Context) {
	// TODO: Implement cost retrieval
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func getCostSummary(c *gin.Context) {
	// TODO: Implement cost summary
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func getCostForecast(c *gin.Context) {
	// TODO: Implement cost forecast
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func analyzeResources(c *gin.Context) {
	// TODO: Implement resource analysis
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func getRecommendations(c *gin.Context) {
	// TODO: Implement recommendations retrieval
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func applyRecommendations(c *gin.Context) {
	// TODO: Implement recommendations application
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func getProviders(c *gin.Context) {
	// TODO: Implement providers list
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func getProviderDetails(c *gin.Context) {
	// TODO: Implement provider details
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func connectProvider(c *gin.Context) {
	// TODO: Implement provider connection
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func disconnectProvider(c *gin.Context) {
	// TODO: Implement provider disconnection
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func getResources(c *gin.Context) {
	// TODO: Implement resources list
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func getResource(c *gin.Context) {
	// TODO: Implement resource details
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func scanResources(c *gin.Context) {
	// TODO: Implement resource scanning
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}

func tagResources(c *gin.Context) {
	// TODO: Implement resource tagging
	c.JSON(http.StatusNotImplemented, gin.H{"error": "Not implemented"})
}